Fix evaluation output duplication
With this way of implementing limited recursion, the bound expressions get evaluated again with every recursive application, which would not be necessary if recursion was not limited.
I can't find a better way to do it for now, because limited recursion implies that the results of evaluation can change (when the limit is reached).
You would somehow need to change the value the evaluation of a bound variable resulted in after the fact.
Other options include limiting all uses of bound expressions (instead of recursive uses) or maybe something with graph-reduction.